src/requirements.txt
Flask==3.0.3
google-cloud-storage==2.16.0
google-api-python-client
google-generativeai
faiss-cpu==1.8.0
sentence-transformers==2.7.0
numpy==1.26.4
sqlparse==0.5.0
pickle5==0.0.11
